Renamed to end effector in robot container, got rid of all setpoint clauses
