+++ src/main/java/frc/robot/commands/elevator/ElevatorPosition.java
@@ -70,6 +70,7 @@ public class ElevatorPosition extends Command {
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return m_pid.atSetpoint();
+    // return m_pid.atSetpoint();
+    return false;
   }
 }
